+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/roots/SqueakContextNode.java
@@ -45,8 +45,8 @@ public class SqueakContextNode extends RootNode {
     }
 
     private static VirtualFrame createFrame(CompiledCodeObject method, ListObject ctxt) {
-        int pc = (int) ctxt.at0(ContextParts.PC.ordinal()).unsafeUnwrapInt();
-        int sp = (int) ctxt.at0(ContextParts.SP.ordinal()).unsafeUnwrapInt();
+        int pc = ctxt.at0(ContextParts.PC.ordinal()).unsafeUnwrapInt();
+        int sp = ctxt.at0(ContextParts.SP.ordinal()).unsafeUnwrapInt();
         BaseSqueakObject closure = ctxt.at0(ContextParts.CLOSURE.ordinal());
         BaseSqueakObject receiver = ctxt.at0(ContextParts.RECEIVER.ordinal());
         VirtualFrame frame = Truffle.getRuntime().createVirtualFrame(EMPTY_ARRAY, method.getFrameDescriptor());
